requirements.txt
python-dotenv==0.10.3
markovify==0.7.1
PyTumblr==0.0.8
